@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
@@ -23,16 +23,14 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
-  bloc: ^6.0.1
-  flutter_bloc: ^6.0.1
+  bloc: ^7.0.0-nullsafety.3
+  flutter_bloc: ^7.0.0-nullsafety.3
   koin:
     path: ../../packages/koin
   koin_flutter: 
     path: ../../packages/koin_flutter
   koin_bloc: 
     path: ../../packages/koin_bloc
-  koin_devtools:
-    path: ../../packages/koin_devtools
 dev_dependencies:
   flutter_test:
     sdk: flutter
