@@ -1,15 +1,15 @@
 name: koin
 description: A pragmatic lightweight dependency injection framework for Dart projects.
-version: 0.14.2
+version: 0.15.0-nullsafety.0
 homepage: https://github.com/pbissonho/koin.dart
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
-  kt_dart: ^0.8.0
-  meta: ^1.1.6
+  kt_dart: ^0.9.0-nullsafety.0
+  meta: ^1.3.0
 dev_dependencies:
   effective_dart: ^1.2.0
-  test: ^1.6.0
-  test_coverage: ^0.3.0
+  test: ^1.16.0-nullsafety.12
+  test_coverage: ^0.5.0
