@@ -1,17 +1,18 @@
 name: koin_bloc
 description: A package to make it easier to use Bloc library with Koin.dart.
-version: 0.1.2
+version: 0.2.0-nullsafety.1
 homepage: https://github.com/pbissonho/koin.dart
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
-  bloc: ^6.0.3
-  koin: ^0.14.0
+  bloc: ^7.0.0
+  koin: ^0.15.0-nullsafety.0
   
 dev_dependencies:
   pedantic: ^1.9.0
-  koin_test: ^0.12.0 
-  test: ^1.14.4
+  koin_test: ^0.13.0-nullsafety.0
+  test: ^1.16.0-nullsafety.12
+  test_coverage: ^0.5.0
     
