@@ -1,20 +1,20 @@
 name: koin_flutter
 description: A package to make it easier to use Koin.dart on the Flutter.
-version: 0.10.0
+version: 0.11.0-nullsafety.0
 homepage: https://github.com/pbissonho/koin.dart
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  koin: ^0.14.0
+  koin: ^0.15.0-nullsafety.0
     
   
 dev_dependencies:
   effective_dart: ^1.2.0
-  koin_test: ^0.12.0
+  koin_test: ^0.13.0-nullsafety.0
   flutter_test:
     sdk: flutter
 
