@@ -14,11 +14,10 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
-  koin:   
-    path: ../../koin  
+  koin: ^0.15.0-nullsafety.0 
   
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -26,7 +25,6 @@ dependencies:
   
 
 dev_dependencies:
-  mockito: any
   koin_test:
     path: ../
   flutter_test:
