@@ -1,19 +1,18 @@
 name: koin_test
 description: A testing library for koin, makes easy mock of definitions and module testing.
-version: 0.12.1
+version: 0.13.0-nullsafety.0
 homepage: https://github.com/pbissonho/koin.dart
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
-  koin: ^0.14.2
-  kt_dart: ^0.8.0
-  equatable: ^1.2.2
-  test: ^1.6.0
-  meta: ^1.1.6
+  koin: ^0.15.0-nullsafety.0
+  kt_dart: ^0.9.0-nullsafety.0
+  test: ^1.16.0-nullsafety.12
+  meta: ^1.3.0
 
 dev_dependencies:
   pedantic: ^1.8.0
-  mockito: ^4.1.1
+  test_coverage: ^0.5.0
  
