@@ -3,4 +3,4 @@ export 'src/definition/provider_definition.dart' show ProviderDefinition;
 export 'src/context/context_handler.dart';
 export 'src/context/context.dart';
 export 'src/definition/parameter.dart';
-export 'src/observer.dart';
+export 'src/observer/observer.dart';
