@@ -5,8 +5,9 @@ typedef ProviderCreate<T> = T Function(Scope scope);
 
 typedef ProviderCreateParam<T, A> = T Function(Scope scope, A param);
 
+// ignore: one_member_abstracts
 abstract class ProviderCreateBase<T> {
-  T create(Parameter parameter, Scope scope);
+  T create(Parameter? parameter, Scope scope);
 }
 
 class ProviderCreateDefinition<T> implements ProviderCreateBase<T> {
@@ -15,7 +16,7 @@ class ProviderCreateDefinition<T> implements ProviderCreateBase<T> {
   ProviderCreateDefinition(this._create);
 
   @override
-  T create(Parameter parameter, Scope scope) {
+  T create(Parameter? parameter, Scope scope) {
     return _create(scope);
   }
 }
@@ -26,7 +27,7 @@ class ProviderCreateParamDefinition<T, A> implements ProviderCreateBase<T> {
   ProviderCreateParamDefinition(this._create);
 
   @override
-  T create(Parameter parameter, Scope scope) {
-    return _create(scope, parameter.get());
+  T create(Parameter? parameter, Scope scope) {
+    return _create(scope, parameter?.get());
   }
 }
