@@ -5,25 +5,26 @@ import 'provider_definition.dart';
 
 class Definitions {
   static ProviderDefinition<T> saveSingle<T>(
-      Qualifier qualifier,
+      Qualifier? qualifier,
       ProviderCreateBase<T> providerCreate,
       ScopeDefinition scopeDefinition,
       Options options) {
-    var beanDefinition = createSingle<T>(
+    final beanDefinition = createSingle<T>(
         qualifier: qualifier,
         providerCreate: providerCreate,
         scopeDefinition: scopeDefinition,
-        options: options);
+        options: options,
+        secondaryTypes: []);
     scopeDefinition.save(beanDefinition);
     return beanDefinition;
   }
 
   static ProviderDefinition<T> createSingle<T>(
-      {Qualifier qualifier,
-      ProviderCreateBase<T> providerCreate,
-      ScopeDefinition scopeDefinition,
-      Options options,
-      List<Type> secondaryTypes}) {
+      {Qualifier? qualifier,
+      required ProviderCreateBase<T> providerCreate,
+      required ScopeDefinition scopeDefinition,
+      required Options options,
+      required List<Type> secondaryTypes}) {
     return ProviderDefinition<T>(
         scopeDefinition: scopeDefinition,
         primaryType: T,
@@ -35,11 +36,11 @@ class Definitions {
   }
 
   static ProviderDefinition<T> createFactory<T>(
-      {Qualifier qualifier,
-      ProviderCreateBase<T> providerCreate,
-      ScopeDefinition scopeDefinition,
-      Options options,
-      List<Type> secondaryTypes}) {
+      {Qualifier? qualifier,
+      required ProviderCreateBase<T> providerCreate,
+      required ScopeDefinition scopeDefinition,
+      required Options options,
+      required List<Type> secondaryTypes}) {
     return ProviderDefinition<T>(
         scopeDefinition: scopeDefinition,
         primaryType: T,
@@ -51,15 +52,16 @@ class Definitions {
   }
 
   static ProviderDefinition<T> saveFactory<T>(
-      Qualifier qualifier,
+      Qualifier? qualifier,
       ProviderCreateBase<T> providerCreate,
       ScopeDefinition scopeDefinition,
       Options options) {
-    var beanDefinition = createFactory<T>(
+    final beanDefinition = createFactory<T>(
         qualifier: qualifier,
         providerCreate: providerCreate,
         scopeDefinition: scopeDefinition,
-        options: options);
+        options: options,
+        secondaryTypes: []);
     scopeDefinition.save(beanDefinition);
     return beanDefinition;
   }
