@@ -18,15 +18,10 @@ import 'measure.dart';
 import 'koin_dart.dart';
 import 'logger.dart';
 import 'module.dart';
+import 'observer/empty_logger_instance_observer.dart';
 
-//
 /// Koin Application
 /// Help prepare resources for Koin context
-// @author - Arnaud Giuliani
-//
-// Ported to Dart from Kotlin by:
-// @author - Pedro Bissonho
-//
 class KoinApplication {
   Koin koin = Koin();
 
@@ -88,8 +83,8 @@ class KoinApplication {
   /// Set Koin Logger
   /// @param logger - logger
   ///
-  KoinApplication emptyInstanceLoggerObserver() {
-    koin.loggerInstanceObserver = null;
+  KoinApplication emptyLoggerObserver() {
+    koin.loggerObserver = EmptyLoggerObserver();
     return this;
   }
 
