@@ -17,57 +17,37 @@
 import 'context/context_handler.dart';
 import 'qualifier.dart';
 import 'koin_dart.dart';
-import 'lazy.dart';
 import 'scope/scope.dart';
 
 ///
 /// KoinComponentMixin interface marker to bring Koin extensions features
 ///
-/// @author Arnaud Giuliani
-///
-
 mixin KoinComponentMixin {
   ///
   /// Get the associated Koin instance
   ///
   Koin getKoin() => KoinContextHandler.get();
 
-  T get<T>([Qualifier qualifier]) {
+  T get<T>([Qualifier? qualifier]) {
     return getKoin().get<T>(qualifier);
   }
 
-  T getWithParam<T, P>(P param, {Qualifier qualifier}) {
+  T getWithParam<T, P>(P param, {Qualifier? qualifier}) {
     return getKoin().getWithParam<T, P>(param);
-  }
-
-  ///
-  /// Lazy inject instance from Koin
-  /// @param qualifier
-  /// @param parameters
-  ///
-  Lazy<T> inject<T>([Qualifier qualifier]) {
-    return getKoin().inject<T>(qualifier);
-  }
-
-  ///
-  /// Lazy inject instance from Koin
-  ///
-  Lazy<T> injectWithParam<T, P>(P param, {Qualifier qualifier}) {
-    return getKoin().injectWithParam<T, P>(param, qualifier: qualifier);
   }
 
   ///
   /// Get instance instance from Koin by Primary Type P, as secondary type S
   /// @param parameters
   ///
-  S bind<S, P>([Qualifier qualifier]) {
+  S bind<S, P>([Qualifier? qualifier]) {
     return getKoin().bind<S, P>();
   }
 
   ///
   /// Get instance instance from Koin by Primary Type K, as secondary type S
   /// @param parameters
-  S bindWithParam<S, K, P>(P param, {Qualifier qualifier}) {
+  S bindWithParam<S, K, P>(P param, {Qualifier? qualifier}) {
     return getKoin().bindWithParam<S, K, P>(param, qualifier: qualifier);
   }
 }
@@ -78,35 +58,19 @@ mixin ScopedComponentMixin {
   ///
   Scope get componentScope;
 
-  T get<T>([Qualifier qualifier]) {
+  T get<T>([Qualifier? qualifier]) {
     return componentScope.get<T>(qualifier);
   }
 
-  T getWithParam<T, P>(P param, {Qualifier qualifier}) {
+  T getWithParam<T, P>(P param, {Qualifier? qualifier}) {
     return componentScope.getWithParam<T, P>(param, qualifier: qualifier);
-  }
-
-  ///
-  /// Lazy inject instance from Koin
-  /// @param qualifier
-  /// @param parameters
-  ///
-  Lazy<T> inject<T>([Qualifier qualifier]) {
-    return componentScope.inject<T>(qualifier);
-  }
-
-  ///
-  /// Lazy inject instance from Koin
-  ///
-  Lazy<T> injectWithParam<T, P>(P param, {Qualifier qualifier}) {
-    return componentScope.injectWithParam<T, P>(param, qualifier: qualifier);
   }
 
   ///componentScope
   /// Get instance instance from Koin by Primary Type P, as secondary type S
   /// @param parameters
   ///
-  S bind<S, P>([Qualifier qualifier]) {
+  S bind<S, P>([Qualifier? qualifier]) {
     return componentScope.bind<S, P>(qualifier);
   }
 
@@ -114,8 +78,7 @@ mixin ScopedComponentMixin {
   /// Get instance instance from Koin by Primary Type K, as secondary type S
   /// @param parameters
   ///
-  /// TODO
-  S bindWithParam<S, K, P>(P param, {Qualifier qualifier});
+  S bindWithParam<S, K, P>(P param, {Qualifier? qualifier});
 }
 
 /// KoinComponent interface marker to bring Koin extensions features.
@@ -123,14 +86,17 @@ mixin ScopedComponentMixin {
 /// to make definition resolution .What we need now is an API
 /// to retrieve our instances outside of the container. That's the goal of Koin
 /// components.
-///## Create a Koin Component
-///To give a class the capacity to use Koin features, we need to *tag it* with
-///`KoinComponent` interface. Let's take an example.
+///
+/// ## Create a Koin Component
+///
+/// To give a class the capacity to use Koin features, we need to *tag it* with
+/// `KoinComponent` interface. Let's take an example.
 /// A module to define MyService instance
-///```dart
-///class MyService {}
-///var myModule = module()..single((s) => MyService());
-///```
+///
+/// ```dart
+/// class MyService {}
+/// var myModule = module()..single((s) => MyService());
+/// ```
 /// we start Koin before using definition.
 /// Initializes koin with `startKoin`:
 /// ```dart
@@ -142,28 +108,28 @@ mixin ScopedComponentMixin {
 ///
 ///    // Create MyComponent instance and inject from Koin container
 ///    MyComponent();
-///}
-///```
+/// }
+/// ```
 ///
 /// Here is how we can write our `MyComponent` to retrieve instances from
 /// Koin container.
 ///
 /// Use `get` & by `inject` to inject MyService instance:
 ///
-///```dart
-///class MyComponent extends View with KoinComponentMixin {
-///  MyService myService;
-///  Lazy<MyService> myServiceLazy;
+/// ```dart
+/// class MyComponent extends View with KoinComponentMixin {
+///   MyService myService;
+///   Lazy<MyService> myServiceLazy;
 ///
-///  MyComponent() {
-///    // lazy inject Koin instance
-///    myServiceLazy = inject<MyService>();
-///    // or
-///    // eager inject Koin instance
-///    myService = get<MyService>();
-///  }
-///}
-///```
+///   MyComponent() {
+///     // lazy inject Koin instance
+///     myServiceLazy = inject<MyService>();
+///     // or
+///     // eager inject Koin instance
+///     myService = get<MyService>();
+///   }
+/// }
+/// ```
 /// ## Unlock the Koin API with KoinComponents
 ///
 /// Once you have tagged your class as `KoinComponent`, you gain access to:
@@ -175,38 +141,49 @@ mixin ScopedComponentMixin {
 /// Koin offers two ways of retrieving instances from the Koin container:
 /// * `Lazy<T> t = inject<T>()` - lazy evaluated delegated instance
 /// * `T t = get<T>()` - eager access for instance
-///```dart
-//  is lazy evaluated
-/// Lazy<MyService> myService = inject();
-/// retrieve directly the instance
-/// MyService myService = get();
-///```
+///
+/// ```dart
+///  is lazy evaluated
+///  Lazy<MyService> myService = inject();
+///  retrieve directly the instance
+///  MyService myService = get();
+/// ```
+///
 /// The `lazy` inject form is better to define property that
 /// need lazy evaluation.
+///
 /// ## Resolving instance from its name
+///
 /// If you need you can specify the following parameter with `get` or `inject`
 /// * `qualifier` - name of the definition (when specified name parameter
 ///  in your definition)
-/// Example of module using definitions names:
-/// ```dart
-/// class ComponentA {}
 ///
-/// class ComponentB {
-///  final ComponentA componentA;
-///	///
-/// KoinComponent interface marker to bring Koin extensions features	///  ComponentB(this.componentA);
-///}
-///	///
-/// final myModule = module()
-///  ..single((s) => ComponentA(), qualifier: named("A"))
-///  ..single((s) => ComponentB(s.get()), qualifier: named("B"));
-///```
+/// Example of module using definitions names:
+///
+/// ```dart
+///  class ComponentA {}
+///
+///  class ComponentB {
+///   final ComponentA componentA;
+///
+///  KoinComponent interface marker to bring Koin extensions features
+///  ComponentB(this.componentA);
+///  }
+///
+///  final myModule = module()
+///   ..single((s) => ComponentA(), qualifier: named("A"))
+///   ..single((s) => ComponentB(s.get()), qualifier: named("B"));
+/// ```
+///
 /// We can make the following resolutions:
+///
 ///```dart
 /// retrieve from given module
 /// var a = get<ComponentA>(named("A"))
 ///```
+///
 /// ## No inject() or get() in your API?
+///
 /// If your are using an API and want to use Koin inside it, just tag
 /// the desired class with `KoinComponent` interface.
 class KoinComponent with KoinComponentMixin {}
