@@ -16,17 +16,7 @@
 
 import 'package:meta/meta.dart';
 
-///
-/// Qualifier
-///
-/// @author - Arnaud GIULIANI
-///
-/// Ported to Dart from Kotlin by:
-/// @author - Pedro Bissonho
-///
-
 /// Help qualify a component
-
 @immutable
 abstract class Qualifier {
   final String _value;
@@ -41,13 +31,13 @@ abstract class Qualifier {
   int get hashCode => _value.hashCode;
 
   @override
-  bool operator ==(other) {
+  bool operator ==(Object other) {
     return other is Qualifier && other.value == value;
   }
 }
 
 /// Give a String qualifier
-Qualifier named<T>([String name]) {
+Qualifier named<T>([String? name]) {
   if (name == null) {
     return TypeQualifier(T);
   }
@@ -55,7 +45,7 @@ Qualifier named<T>([String name]) {
 }
 
 /// Give a String qualifier
-Qualifier qualifier<T>([String name]) {
+Qualifier qualifier<T>([String? name]) {
   if (name == null) {
     return TypeQualifier(T);
   }
