@@ -132,7 +132,7 @@ Not can declare and override a single on the fly when override is set to false''
     var a = ComponentA();
 
     var session1 = koin.createScopeWithQualifier('session1', named('Session'));
-    session1.declare(a);
+    session1.declare(a, secondaryTypes: []);
 
     expect(a, session1.get<ComponentA>());
     expect(a, session1.get<ComponentB>().a);
@@ -147,7 +147,7 @@ Not can declare and override a single on the fly when override is set to false''
     var a = ComponentA();
 
     var session1 = koin.createScopeWithQualifier('session1', named('Session'));
-    session1.declare(a);
+    session1.declare(a, secondaryTypes: []);
 
     expect(
         () => koin.get<ComponentA>(), throwsA(isA<NoBeanDefFoundException>()));
@@ -164,7 +164,7 @@ Not can declare and override a single on the fly when override is set to false''
     var session1 = koin.createScopeWithQualifier('session1', named('Session'));
     var session2 = koin.createScopeWithQualifier('session2', named('Session'));
 
-    session1.declare(a);
+    session1.declare(a, secondaryTypes: []);
 
     expect(() => session2.get<ComponentA>(),
         throwsA(isA<NoBeanDefFoundException>()));
