@@ -11,13 +11,13 @@ class CounterCubit extends Cubit<int> {
 class MyScope {}
 
 // Define a module with your cubit
-var cubitModule = Module()
+final cubitModule = Module()
   // Define a single cubit.
   // Single Cubit will be closed when the global context of the koin is closed.
-  ..cubit((s) => CounterCubit())
+  ..bloc((s) => CounterCubit())
   // Define a scoped cubit for MyScopeWidget.
   // Scoped Cubit will be closed when the scope instance is closed,
   //that is, when MyScopeWidget is removed from the widget tree
   ..scope<MyScope>((scope) {
-    scope.scopedCubit<CounterCubit>((scope) => CounterCubit());
+    scope.scopedBloc<CounterCubit>((scope) => CounterCubit());
   });
