@@ -1,7 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:koin/koin.dart';
 import 'package:koin_test/koin_test.dart';
-import 'package:mockito/mockito.dart';
 
 // Just a few services to exemplify.
 class ServiceA {
@@ -30,9 +29,6 @@ class ServiceC {
 
   String getName() => '$firstName $lastName';
 }
-
-// Create a Mock for a particular service as usual.
-class ServiceAMock extends Mock implements ServiceA {}
 
 // Create a Fake for a particular service as usual.
 class ServiceFake extends Fake implements ServiceA {
@@ -68,6 +64,7 @@ void main() {
     }, throwsException);
   });
 
+  /*
   test('shoud return mock instance', () {
     declareModule((module) {
       module
@@ -87,7 +84,7 @@ void main() {
     expect(service, isNotNull);
     expect(service, isA<ServiceAMock>());
     expect(service.getName(), 'MockName');
-  });
+  });*/
 
   test(('shoud return a Fake instance'), () {
     declareModule((module) {
